Добавил делегат для QuestionFactory по инструкции в уроке и все работает
