+++ MovieQuiz/SceneDelegate.swift
@@ -1,3 +1,4 @@
+
 import UIKit
 
 final class SceneDelegate: UIResponder, UIWindowSceneDelegate {
