+++ MovieQuiz/Services/QuestionFactoryDelegate.swift
@@ -1,0 +1,6 @@
+
+import UIKit
+
+protocol QuestionFactoryDelegate: AnyObject {               
+    func didReceiveNextQuestion(question: QuizQuestion?)
+}
