+++ MovieQuiz/Services/QuestionFactoryProtocol.swift
@@ -1,0 +1,7 @@
+
+import UIKit
+
+protocol QuestionFactoryProtocol {
+    func requestNextQuestion()
+}
+
